Mass mailing model and action on Users table
An action 'send mass mail' is implemented on the
table of Users as a redirect on creation of new
mass mail. Mass mail uses the M2M relationship
onto table of Users to create new records in
the table of Communications. This needs to be done
via overriding save_formset() to be able to deal
with the data from the M2M tables.