Allow attachments in Communications and MassCommunications
This does not add attachments to AutomaticCommunications.
It is not clear if and in what form it would be useful, e.g.
for sending tax confirmations. It should probably allow
to pass a generating function instead of fixed attachment
contents. We do not need it now, so we do not worry about
it in this patch.