One time payment forms
Forms to create a new user in the club and a new expected payment. It
is however necessary to deal with the situation that the user might
already have an account, thus the registration form is splitted and if
matches are found in name or email against already registered users, the
wizard tries to verify the user against the existing account.

This requires Django 1.4 alpha for SessionFormWizard