Add special variable action into Conditions
'action' variable is not Null only when the condition is evaluated as
a part of a hook of some action in the program. The value of 'action'
then describes the specific event, such as 'new-user' or
'new-payment'.

This makes it possible to write dynamic conditions (and corresponding
automatic communications) that only trigger for certain events.